# Optimize AGB split functions