--- a/R/Boudewyn_cumPoolsCreateAGB.R
+++ b/R/Boudewyn_cumPoolsCreateAGB.R
@@ -36,7 +36,7 @@ cumPoolsCreateAGB <- function(allInfoAGBin, table6, table7, pixGroupCol){
   if (any(!(expectedColumns %in% colnames(allInfoAGBin)))) {
     stop("The AGB table needs the following columns ", paste(expectedColumns, collapse = " "))
   }
-  AGB <- as.data.table(allInfoAGBin)
+  AGB <- as.data.table(allInfoAGBin, key = NULL)
 
   # 2. Get parameters for all curves
   # Identify all unique species/location combinations
@@ -100,7 +100,6 @@ cumPoolsCreateAGB <- function(allInfoAGBin, table6, table7, pixGroupCol){
 convertAGB2pools <- function(AGB, params6, params7){
   params6 <- merge(AGB, params6, all.x = TRUE)
   params7 <- merge(AGB, params7, all.x = TRUE)
-
   # get the proportions of each pool
   pVect <- biomProp(table6 = params6, table7 = params7, x = AGB$B, type = "biomass")
   totTree <-  AGB$B