I re-wrote the functions to split total above ground biomass to pools. The idea was to get rid of the of the `for` loop by a vectorized approach, and to make use of the efficiency of `data.table`. 

On the toy dataset I created, the new function was taking 1% (!!!) of the time of the previous version. However, when I compared on a real case study (entire RIA), the improvements were much less impressive:

> **Pre change:**
> split yield tables: 5650 ms; +9047.5 -6575.4 MB
> split cohort Data (2x): 9900 ms; +13919.6 -16063.7 MB
> 
> **Post change:**
> split yield tables: 4440 ms; +6899.8 -3460.1 MB
> split cohort Data (2x): 5610 ms; +10687.5 -3646.7 MB

I think that the structure of the data cause the changes to be much less effective (i.e., there are orders of magnitude less "curves" than biomass to split per curve). Still, splitting the cohort data took almost half the time, and considering it is an operation that occurs every timestep, it is a good improvement.

I had to created the function `biomPropAGB` because I did not want to touch the `biomProp` function which is used by `CBM_vol2biomass`. I don't think it would be that much work to apply similar changes to the functions used in `CBM_vol2biomass`. I suspect the improvements to be more marginal though since there are much less Boudewyn "curves" (combination of species and spatial unit) to loop through.