@suz-estella 2025-07-09
> Tests are failing because it could not resolve dependencies for the R packages. @camillegiuliano or @suz-estella, can we check why? is this a case where I should accept anyway because of all the changes we are doing?

Definitely can ignore the failure on windows-latest (release) about not being able to resolve dependency SpaDES.tools. It's on our radar but not affecting our ability to use CBMutils on Windows for now.